+++ components/Chat/index.ts
@@ -1,3 +1,3 @@
 export { default as Chat } from './Chat.component'
 export { useChatContext, ChatContext } from './context'
-export { SideBar } from './components/SideBar.component'
+export { SideBarChatList } from './components/SideBarChatList.component'
