+++ response.json
@@ -1,0 +1,1 @@
+{"id":"chatcmpl-9PHK1TjOolNFTfMLOdnl5a4nTULRx","object":"chat.completion","created":1715811813,"model":"gpt-4-1106-preview","choices":[{"index":0,"message":{"role":"assistant","content":"feat: Refactor Chat Pagе and Context Management\n\n### Summary\n\n**Type:** Feature\n\n* **What kind of change does this PR introduce?**\nThis PR introduces a major refactoring of the chat page and the context management system in the application. It simplifies the rendering of the chat page by creating a new `ChatPage` component with cleaner context handling and encapsulating context providers in a new `Contexts` component.\n\n* **What is the current behavior?**\nPreviously, the chat page had a deeply nested structure with multiple contexts intertwined and the context providers (for authentication, chat, and persona) spread across the component hierarchy. Additionally, functionality related to personas was present but spread across various components.\n\n* **What is the new behavior (if this is a feature change)?**\nThe new behavior introduces a simplified chat page, a dedicated `Contexts` component that wraps the necessary context providers, and a `Wrapper` component that includes common UI elements like banner, header, and toaster. The side bar components have been renamed and streamlined, with persona-specific components being removed. The codebase organization is improved by relocating several utility functions to a more appropriate directory. Some default personas have been defined inside the chat context instead of being managed by a persona context.\n\n* **Does this PR introduce a breaking change?**\nThis update includes significant changes, which may require updates to how existing components interact with the refactored chat page and context providers. Implementations that assume the previous structure will need to be revisited.\n\n* **Has Testing been included for this PR?**\nTesting details have not been included and should be provided by the development team.\n\n### Other Information\n\nThe refactoring effort emphasizes better modularization, cleaner context management, and a more streamlined chat page UI. This should enhance maintainability and scalability of the chat feature within the application."},"logprobs":null,"finish_reason":"stop"}],"usage":{"prompt_tokens":9280,"completion_tokens":372,"total_tokens":9652},"system_fingerprint":null}
